+++ System_NES/Mappers/Mapper000.cs
@@ -17,7 +17,7 @@
             // Nothing to do
         }
 
-        public override bool MapCPUAddressRead(ushort address, out uint newAddress, out byte data)
+        public override bool MapCPUAddressRead(uint address, out uint newAddress, out byte data)
         {
             data = 0;
             if (address >= 0x8000 && address <= 0xFFFF)
@@ -30,7 +30,7 @@
             return false;
         }
 
-        public override bool MapCPUAddressWrite(ushort address, byte data, out uint newAddress)
+        public override bool MapCPUAddressWrite(uint address, byte data, out uint newAddress)
         {
             if (address >= 0x8000 && address <= 0xFFFF)
             {
